Cache allows script to quit early
- cache will quit early if already downloaded
- displays [ downloaded / filtered ] counts
- less .delay() if file exists

+++ srid.py
@@ -2,13 +2,35 @@ from __future__ import print_function, division
 import os
 import time
 import urllib
+import pickle
 
 import praw
 
-SLEEP_IN_SECS = 2.  # 0.3
-version = "0.2.2"
+SLEEP_IN_SECS = 1.3  # 0.3
+version = "0.2.3"
 debug = False
 stats_downloaded = 0
+stats_filtered = 0
+
+
+def load_cache():
+    # get last-retrieved
+    path = os.path.join("srid-downloaded", "cache.pkl")
+
+    try:
+        with open(path, "rb") as f:
+            cache = pickle.load(f)
+            return cache
+
+    except IOError:
+        return (None, None, None)
+
+
+def save_cache(cache):
+    # save last-retrieved
+    path = os.path.join("srid-downloaded", "cache.pkl")
+    with open(path, "wb") as f:
+        pickle.dump(cache, f)
 
 
 def sanitize_filename(filename):
@@ -23,7 +45,7 @@ def sanitize_filename(filename):
 
 def download_file(url, subreddit, title):
     # save from url, to ./srid-downloaded/{subreddit}/{title}
-    global stats_downloaded
+    global stats_downloaded, stats_filtered
     ext = os.path.splitext(url)[1]
 
     # imgur has bad file names
@@ -41,8 +63,13 @@ def download_file(url, subreddit, title):
     if debug:
         print("writing...: ", filename)
 
-    urllib.urlretrieve(url, filename)
-    stats_downloaded += 1
+    # delay if download, count stats
+    if not os.path.exists(filename):
+        stats_downloaded += 1
+        urllib.urlretrieve(url, filename)
+        time.sleep(SLEEP_IN_SECS)
+
+    stats_filtered += 1
 
 
 def do_filter(likes, subs):
@@ -50,18 +77,30 @@ def do_filter(likes, subs):
     # Should I instead should delete or use that built in func/class to
     # filter+delete ?
 
+    # load prev cache, save new one
+    cache = load_cache()
+    cached = False
+
+    print("cache: ", cache)
     print("Downloading: ")
+
     for cur in likes:
+        # save first item downloaded. But I want to keep the generator.
+        if not cached:
+            save_cache((cur.url, cur.title, cur.subreddit.url))
+            cached = True
+
+        # can I quit early?
+        if cache == (cur.url, cur.title, cur.subreddit.url):
+            print("updated to cached, ending.")
+            return
+
+        # if any sub will match.
         if any(sub.lower() in cur.subreddit.url.lower() for sub in subs):
-            #if subs[0].lower() in cur.subreddit.url.lower():
             if debug:
                 print("get: ", cur.url, " -- ", cur.subreddit.url)
             download_file(cur.url, cur.subreddit.display_name, cur.title)
             print(".", end="")
-            time.sleep(SLEEP_IN_SECS)
-        else:
-            pass
-            #print("skips: ", cur.subreddit.url, cur.url)
 
 
 def print_debug(likes):
@@ -89,23 +128,19 @@ def main(sub_names, num=10):
     if debug:
         print("logged in? ", isinstance(r.user, praw.objects.LoggedInRedditor))
 
-    # == list form ==
-    #likes = list(r.user.get_liked(num))
-    #if debug: print_debug(likes)
-    #do_filter(likes, sub_names)
-
-    # == gen form ==, will create on-demand so it should be easier/quicker
+    # .get_liked() is a generator, will create on-demand
     do_filter(r.user.get_liked(limit=num), sub_names)
 
 if __name__ == "__main__":
     # hardcoded for now
     subs_photos = ["EarthPorn", "AnimalPorn", "SpacePorn", "wallpaper", "itookapicture", "photocritique"]
-    subs_imaginary = ["imaginaryArmor", "imaginaryBattlefields", "imaginaryCharacters", "imaginaryCityscapes", "imaginaryLandscapes", "imaginaryRobotics", "imaginaryStarships", "imaginaryTechnology", "imaginaryVehicles", "imaginaryWeaponry", "imaginaryMonsters", "SpecArt"]
-    subs_sfw_images = ["microporn", "AbandonedPorn", "futurePorn", "AdPorn", "AlbumArtPorn", "ArchitecturePorn", "ArtPorn", "BookPorn", "CityPorn", "DesignPorn", "DestructionPorn", "EarthPorn", "FirePorn", "HistoryPorn", "GeologyPorn", "images", "FossilPorn", "MilitaryPorn", "MapPorn", "MoviePosterPorn", "ColorizedHistory", "BattlePaintings", "FighterJets", "Airplanes", "Helicopters", "WarshipPorn", "Helicopters", "HumanPorn", "OldSchoolCool", "TheWayWeWere", "VintageAds", "PropagandaPosters", "Castles", "concertposterporn", "VHScoverART", "geekporn", "waterporn", "quotesporn", "ruralporn", "macroporn", "winterporn"]
+    subs_imaginary = ["imaginary", "imaginaryArmor", "imaginaryBattlefields", "imaginaryCharacters", "imaginaryCityscapes", "imaginaryLandscapes", "imaginaryRobotics", "imaginaryStarships", "imaginaryTechnology", "imaginaryVehicles", "imaginaryWeaponry", "imaginaryMonsters", "SpecArt", "imaginaryLandscapes"]
+    subs_sfw_images = ["microporn", "AbandonedPorn", "futurePorn", "AdPorn", "AlbumArtPorn", "ArchitecturePorn", "ArtPorn", "BookPorn", "CityPorn", "DesignPorn", "DestructionPorn", "EarthPorn", "FirePorn", "HistoryPorn", "GeologyPorn", "images", "FossilPorn", "MilitaryPorn", "MapPorn", "MoviePosterPorn", "ColorizedHistory", "BattlePaintings", "FighterJets", "Airplanes", "Helicopters", "WarshipPorn", "Helicopters", "HumanPorn", "OldSchoolCool", "TheWayWeWere", "VintageAds", "PropagandaPosters", "Castles", "concertposterporn", "VHScoverART", "geekporn", "waterporn", "quotesporn", "ruralporn", "macroporn", "winterporn", "ArchitecturePorn"]
     subs_people = ["PrettyGirls", "Goddesses", "FineLadies", "gentlemanboners", "ClassicScreenBeauties", "ladyladyboners", "ladyboners", "VGB", "VintageLadyBoners", "faces", "classywomenofcolor"]
 
     subs = subs_photos + subs_imaginary + subs_sfw_images + subs_people
 
-    print("downloading: ")
+    print("downloading: sleep = {}secs".format(SLEEP_IN_SECS))
     main(subs, num=100)
-    print("\nDownloaded: ", stats_downloaded)
+    print("\nDownloaded: {} [filtered: {}], ".format(
+            stats_downloaded, stats_filtered))
